+++ Njsast/Bundler/ImportExportTransformer.cs
@@ -85,6 +85,22 @@ namespace Njsast.Bundler
                         return Remove;
                     }
 
+                    if (call.Expression.IsSymbolDef().IsGlobalSymbol() == "__createBinding" && call.Args.Count >= 3 &&
+                        call.Args[0].IsSymbolDef().IsExportsSymbol() && call.Args[2] is AstString bindingName &&
+                        DetectImport(call.Args[1]) is {} bindModule && bindModule.Item2.Length == 0)
+                    {
+                        if (call.Args.Count == 3)
+                        {
+                            _sourceFile.SelfExports.Add(new ReexportSelfExport(bindingName.Value, bindModule.Item1, Concat(bindModule.Item2, bindingName.Value)));
+                            return Remove;
+                        }
+                        if (call.Args.Count == 4 && call.Args[3] is AstString asName)
+                        {
+                            _sourceFile.SelfExports.Add(new ReexportSelfExport(asName.Value, bindModule.Item1, Concat(bindModule.Item2, bindingName.Value)));
+                            return Remove;
+                        }
+                    }
+
                     var callSymbol = call.Expression.IsSymbolDef();
                     if ((callSymbol == _reexportSymbol || callSymbol.IsTsReexportSymbol()) && call.Args.Count >= 1)
                     {
@@ -180,13 +196,14 @@ namespace Njsast.Bundler
                     }
                     else
                     {
-                        var trueValue = Transform(pea.Value.value);
                         var newVar = new AstVar(stmBody);
                         var astSymbolVar = new AstSymbolVar(stmBody, newName);
-                        astSymbolVar.Thedef = new SymbolDef(_sourceFile.Ast, astSymbolVar, trueValue);
+                        astSymbolVar.Thedef = new SymbolDef(_sourceFile.Ast, astSymbolVar, null);
                         _sourceFile.Ast.Variables!.Add(newName, astSymbolVar.Thedef);
-                        newVar.Definitions.Add(new AstVarDef(astSymbolVar, trueValue));
                         _exportName2VarNameMap[pea.Value.name] = astSymbolVar.Thedef;
+                        var trueValue = Transform(pea.Value.value);
+                        newVar.Definitions.Add(new AstVarDef(astSymbolVar, trueValue));
+                        astSymbolVar.Thedef.Init = trueValue;
                         _sourceFile.SelfExports.Add(new SimpleSelfExport(pea.Value.name,
                             new AstSymbolRef(_sourceFile.Ast, astSymbolVar.Thedef, SymbolUsage.Unknown)));
                         return newVar;