+++ Test/Input/Bundler/BundleBug5/out/cb-bundle.js
@@ -1,0 +1,8 @@
+(function(undefined) {
+    "use strict";
+    function a_lib2() {
+        console.log("ok");
+    }
+    a_lib2();
+}).call(this);
+