+++ Test/Input/Bundler/BundleBug5/out/cbm-bundle.js
@@ -1,0 +1,8 @@
+(function(n) {
+    "use strict";
+    function i() {
+        console.log("ok");
+    }
+    i();
+}).call(this);
+