+++ Test/Input/Bundler/BundleBug5/out/cm-bundle.js
@@ -1,0 +1,1 @@
+(function(n){"use strict";function i(){console.log("ok")}i()}).call(this)